requirements.txt
google-genai>=0.5.0
python-dotenv>=1.0.0
rich>=13.0.0
requests>=2.28.0
python-telegram-bot>=20.0
PyPDF2>=3.0.0
python-docx>=0.8.11
pytest>=7.0.0
playwright>=1.40.0
beautifulsoup4>=4.12.0
prompt_toolkit>=3.0.0
